+++ src/SBMoffat.cpp
@@ -25,7 +25,6 @@
 #include "SBMoffatImpl.h"
 #include "integ/Int.h"
 #include "Solve.h"
-#include "math/BesselRoots.h"
 #include "math/Bessel.h"
 #include "math/Gamma.h"
 #include "math/Angle.h"