+++ src/math/Hankel.cpp
@@ -27,7 +27,6 @@
 #include "integ/Int.h"
 #include "math/Hankel.h"
 #include "math/Bessel.h"
-#include "math/BesselRoots.h"
 #include "Std.h"
 
 namespace galsim {